requirements.txt
amqp==5.0.6
asgiref==3.4.1
autopep8==1.5.7
billiard==3.6.4.0
celery==5.2.1
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.7
django-celery-beat==2.2.1
django-cors-headers==3.10.0
django-modeltranslation==0.17.3
django-rosetta==0.9.7
django-timezone-field==4.2.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
idna==3.3
kombu==5.2.2
Pillow==8.3.2
polib==1.1.1
prompt-toolkit==3.0.23
psycopg2-binary==2.9.1
pycodestyle==2.7.0
PyJWT==2.3.0
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2021.3
redis==3.5.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
gunicorn==20.1.0